feature/1: Implemented client entity mixin

+++ src/main/java/com/blukzen/createlab/util/IEntityMixin.java
@@ -6,5 +6,5 @@ public interface IEntityMixin {
     boolean isInsideLabPortal();
     boolean isOnLabPortalCooldown();
     void setLabPortalCooldown();
-    int getLabPortalCooldown();
+    float getLabPortalCooldown();
 }
